% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estima_CORR_peso.R
\name{estima_CORR_peso}
\alias{estima_CORR_peso}
\title{estima_CORR_peso}
\usage{
estima_CORR_peso(dados, variaveis, metodo = "pearson")
}
\arguments{
\item{dados}{Um data frame contendo as variáveis de interesse e a coluna de pesos.}

\item{variaveis}{Uma lista de pares de variáveis para as quais a correlação ponderada será calculada.
Cada elemento da lista é uma sublista com dois vetores, onde o primeiro vetor contém as variáveis para x,
e o segundo vetor contém as variáveis para y.}

\item{metodo}{Um string indicando o método de correlação a ser usado, sendo o padrão "pearson". Outros métodos suportados
pela função wdm::indep_test incluem "spearman" e "kendall".}
}
\value{
Retorna um `tibble` com as seguintes colunas:
\describe{
  \item{var1}{Nome do primeiro grupo de variáveis.}
  \item{var2}{Nome do segundo grupo de variáveis.}
  \item{Correlacao}{Valor da correlação estimada.}
  \item{statistic}{Estatística do teste de correlação.}
  \item{p_valor}{Valor-p do teste de correlação.}
  \item{sig_stars}{Indicação da significância do valor-p com símbolos "*".}
  \item{Metodo}{Método de correlação utilizado.}
  \item{hip_alternativa}{Hipótese alternativa do teste.}
  \item{var1_vars}{Lista de variáveis agrupadas no `x`.}
  \item{var2_vars}{Lista de variáveis agrupadas no `y`.}
}
}
\description{
Esta função interna calcula a correlação ponderada entre pares de variáveis especificadas,
aplicando o teste de independência para cada par utilizando o método definido.
}
\details{
Esta função realiza o teste de independência para cada par de variáveis fornecidas, ponderando pelos valores da coluna `peso`.
As variáveis `x` e `y` são representadas pela média das variáveis contidas nos vetores de entrada.
A função aplica correlação ponderada para os pares de variáveis e retorna uma tabela com as estatísticas dos testes,
incluindo valores-p e uma representação da significância.
}
\keyword{internal}
