% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndiceCalculoImportancia.R
\name{IndiceCalculoImportancia}
\alias{IndiceCalculoImportancia}
\title{IndiceCalculoImportancia}
\usage{
IndiceCalculoImportancia(
  dados_split,
  all_importancia,
  referencia,
  remover_quem_n_tem_imp_area_nota = TRUE
)
}
\arguments{
\item{dados_split}{Lista de data frames, onde cada elemento representa um conjunto particionado de dados
de uma distribuidora específica, com informações sobre índices e atributos.}

\item{all_importancia}{Vetor de strings indicando os nomes de índices que devem ser considerados no cálculo da importância.}

\item{referencia}{Data frame de referência que contém as definições e atributos de cada índice, incluindo configurações
de fatores de correção e critérios de inclusão.}

\item{remover_quem_n_tem_imp_area_nota}{Se for calcular apenas importância da área, deixar como FALSE; caso contrário, como TRUE.}
}
\value{
Retorna uma lista com quatro data frames:
  - `importancia_nota`: Notas médias calculadas para cada atributo e distribuidora.
  - `importancia_peso`: Importância relativa calculada com base nos pesos atribuídos a cada atributo.
  - `importancia_q5_pc`: Cálculo ponderado e ajustado da importância por atributo.
  - `qualidade_preco`: Dados de correlação entre qualidade e preço, por distribuidora.
}
\description{
Realiza o cálculo da importância de índices em dados particionados, aplicando transformações, filtros e ponderações
para calcular importância relativa, notas médias, e correlações com atributos de qualidade e preço.
}
\details{
A função `IndiceCalculoImportancia` divide os dados por distribuidora, aplica um filtro de condições de cálculo e calcula índices de importância, notas, e
atributos relativos a cada distribuição. O cálculo inclui a avaliação dos atributos com base em ponderações configuradas,
aplicando ajustes específicos para tipos de índice, como \code{"ISQP"} e \code{"ISCAL"}. Adicionalmente, calcula a correlação entre qualidade e preço
conforme especificado no parâmetro de referência. A função retorna os resultados como uma lista com quatro componentes.
}
\section{Etapas de Cálculo}{

- **Configuração Inicial e Filtros**: Gera filtros de formatação e identifica as distribuidoras e atributos aplicáveis.
- **Ordens e Notas**: Inverte e pondera a ordem dos atributos e calcula notas médias para cada atributo avaliado.
- **Correção de Peso e Cálculo Final**: Aplica fatores de correção, calcula a importância ponderada e percentual, e organiza as saídas por distribuidora.
- **Correlação Qualidade-Preço**: Calcula métricas de correlação de qualidade e preço para os atributos especificados.
}

\examples{

base::print("Sem Exemplo")

}
