% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndiceCalculoIndices.R
\name{IndiceCalculoIndices}
\alias{IndiceCalculoIndices}
\title{IndiceCalculoIndices}
\usage{
IndiceCalculoIndices(
  splits,
  vou_rodar,
  referencia_indices,
  dados,
  amostra_pretendida = NULL,
  all_indices,
  qualidade_preco,
  importancia_q5_pc,
  fonte_imp_entra_serie_hist,
  caminho_temporario = "99.temp cal indi",
  caminho_resultado = "Output",
  remover_dos_indices = c("ISQP", "IEQP", "IIQP", "ISCP", "ISC", "IESC", "IIC", "ISCAL",
    "IECP", "IICP"),
  o_que_remover = c("media", "n", "n_peso", "linhas", "linhas_peso", "1", "2", "3", "4",
    "5", "6", "7", "8", "9", "10", "11", "12", "indiceescala_nota1",
    "indiceescala_nota2", "indiceescala_nota3", "indiceescala_nota4",
    "indiceescala_nota5", "indiceescala_nota6", "indiceescala_nota7",
    "indiceescala_nota8", "indiceescala_nota9", "indiceescala_nota10",
    "indiceescala_nota11", "indiceescala_nota12"),
  nome_salvar = "Indice"
)
}
\arguments{
\item{splits}{Um `data.frame` que contém os divisores e indicadores específicos para o cálculo dos índices.}

\item{vou_rodar}{Um `character` indicando o grupo ou divisão a ser processado, exemplo `c("DISTRIBUIDORA1", "DISTRIBUIDORA2")`.}

\item{referencia_indices}{Um `data.frame` de referência dos índices, que inclui informações sobre as variáveis
de cada índice, necessidade de ponderação e parâmetros adicionais.}

\item{dados}{Um `data.frame` com os dados a serem utilizados no cálculo dos índices.}

\item{amostra_pretendida}{Um `data.frame` opcional que especifica a amostra alvo para o cálculo dos índices.}

\item{all_indices}{Um `character vector` contendo os identificadores dos índices que serão calculados.}

\item{qualidade_preco}{Um `data.frame` com as ponderações relativas à qualidade e preço para os indicadores.}

\item{importancia_q5_pc}{Um `data.frame` com as importâncias relativas dos atributos de cada índice.}

\item{fonte_imp_entra_serie_hist}{Um `character` que define a fonte da importância a ser usada para a série histórica.}

\item{caminho_temporario}{Um `character` que especifica o diretório temporário onde os resultados intermediários
serão salvos, padrão `"99.temp cal indi"`.}

\item{caminho_resultado}{Um `character` que define o caminho onde o resultado final será salvo, padrão `"Output"`.}

\item{remover_dos_indices}{Um `character vector` indicando quais índices devem ser removidos da análise, padrão inclui vários identificadores como `"ISQP", "IEQP"`.}

\item{o_que_remover}{Um `character vector` especificando quais variáveis devem ser removidas dos índices durante a análise, padrão inclui variáveis como `"media", "n", "linhas"`.}

\item{nome_salvar}{Um `character` que define o nome do arquivo de saída, padrão `"Índice"`.}
}
\description{
Esta função calcula índices de satisfação e importância com base em uma série de dados e parâmetros de referência.
Realiza uma análise detalhada de indicadores como médias ponderadas e métricas de NPS (Net Promoter Score),
aplicando ponderação de qualidade e preço conforme especificado.
}
\details{
A função executa uma série de cálculos para avaliar o desempenho de índices de satisfação. As principais etapas incluem:
- Filtragem e pré-processamento dos dados, incluindo o ajuste de ponderações de qualidade e preço.
- Cálculo de médias ponderadas e índices específicos, considerando NPS e critérios de importância relativa.
- Geração de um arquivo temporário com os resultados intermediários para cada divisão ou indicador.
- Salvamento dos resultados finais no diretório de saída especificado.
}
\examples{

base::print("Sem Exemplo")

}
