% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicePlanilhaCorrelacao.R
\name{IndicePlanilhaCorrelacao}
\alias{IndicePlanilhaCorrelacao}
\title{IndicePlanilhaCorrelacao}
\usage{
IndicePlanilhaCorrelacao(
  wb,
  logo_innovare,
  nome_aba,
  ano_titulo_planilha,
  pasta_nome,
  vetor_titulo,
  indice_sigla_preco = "PR1",
  nota_roda_pe = NULL,
  espanhol = FALSE,
  dados,
  Tab_Fundo_cinza,
  referencia_filtrado,
  cor_fundo_menu,
  formatacao_numero = "0.0",
  print = FALSE
)
}
\arguments{
\item{wb}{Workbook para adicionar a nova aba.}

\item{logo_innovare}{Caminho para o logo a ser inserido.}

\item{nome_aba}{Nome da aba a ser criada.}

\item{ano_titulo_planilha}{Ano exibido no título da planilha.}

\item{pasta_nome}{Nome da pasta para exibição no cabeçalho.}

\item{vetor_titulo}{Vetor com títulos das colunas.}

\item{indice_sigla_preco}{Sigla a ser substituída para "PR" nos dados de texto.}

\item{nota_roda_pe}{Vetor de notas de rodapé (opcional).}

\item{espanhol}{Lógico para definir idioma do cabeçalho (padrão: FALSE).}

\item{dados}{Data frame com os dados da planilha.}

\item{Tab_Fundo_cinza}{Data frame com linhas que devem ter fundo cinza.}

\item{referencia_filtrado}{Data frame com cores específicas de indicadores.}

\item{cor_fundo_menu}{Data frame com cores e estilos para cada título de coluna.}

\item{formatacao_numero}{Formatação que entra nos número na 'createStyle'. Por padrão, "0.0".}

\item{print}{Se TRUE, abre a planilha automaticamente após a criação.}
}
\value{
Retorna o workbook `wb` com a nova aba formatada.
}
\description{
Cria uma aba em um `workbook` com formatação visual específica para apresentar correlações entre atributos e satisfação geral, usando cores, mesclagem de células e outras configurações de estilo.
}
\details{
A função configura uma aba completa, mesclando células e aplicando estilos específicos para formatação de correlações. Adiciona cabeçalhos customizados, cores, e estilo condicional de acordo com a natureza dos dados.
}
\examples{

base::print("Sem Exemplo")

}
