% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicePlanilhaEscalaAberta.R
\name{IndicePlanilhaEscalaAberta}
\alias{IndicePlanilhaEscalaAberta}
\title{IndicePlanilhaEscalaAberta}
\usage{
IndicePlanilhaEscalaAberta(
  wb,
  logo_innovare,
  nome_aba,
  pasta_nome,
  manter_ao_remover_dados = NULL,
  remover_dos_indices = NULL,
  indice_sigla_preco = "PR1",
  nota_roda_pe = NULL,
  espanhol = FALSE,
  anos,
  dados,
  Tab_Fundo_cinza,
  referencia_filtrado,
  print = FALSE
)
}
\arguments{
\item{wb}{Workbook onde a aba será criada.}

\item{logo_innovare}{Caminho para a imagem do logotipo a ser inserido na planilha.}

\item{nome_aba}{Nome da aba a ser adicionada ao workbook.}

\item{pasta_nome}{Texto para indicar a pasta de dados no cabeçalho.}

\item{manter_ao_remover_dados}{Colunas a serem preservadas ao remover dados de índices específicos.}

\item{remover_dos_indices}{Vetor de índices que devem ser removidos da planilha.}

\item{indice_sigla_preco}{Sigla do índice de preço para identificar e estilizar linhas específicas.}

\item{nota_roda_pe}{Vetor de notas de rodapé opcionais.}

\item{espanhol}{Lógico. Se TRUE, exibe cabeçalho e títulos em espanhol.}

\item{anos}{Vetor contendo os anos a serem exibidos nos cabeçalhos das colunas.}

\item{dados}{Data frame com os dados que serão exibidos na planilha.}

\item{Tab_Fundo_cinza}{Data frame com informações de cores para formatar linhas específicas.}

\item{referencia_filtrado}{Data frame que define as cores de fundo e texto para cada índice.}

\item{print}{Lógico. Se TRUE, abre a planilha após criação.}
}
\value{
O workbook `wb` atualizado com a aba recém-criada e formatada.
}
\description{
Gera uma planilha formatada com uma escala de satisfação, exibindo dados numéricos e de texto, além de um cabeçalho estilizado. A função monta uma aba nova com organização e estilo específicos de acordo com os parâmetros fornecidos.
}
\details{
Esta função cria uma nova aba no workbook `wb` utilizando `openxlsx`. O layout é estruturado para exibir uma escala aberta de satisfação (5 pontos) e aplica estilos visuais avançados, incluindo coloração por indicadores e mesclagem de células.
Cabeçalhos e títulos são configurados com base no idioma selecionado e no vetor de anos. Estilos adicionais são aplicados a cada célula com base nas cores dos indicadores e configurações de alinhamento e formatação numérica.

A função oferece flexibilidade para ajustar colunas específicas a serem mantidas ou removidas, personalizar o cabeçalho em diferentes idiomas e inserir logotipos. Notas de rodapé opcionais também são suportadas para exibição adicional abaixo da tabela.
}
\examples{

base::print("Sem Exemplo")

}
