% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicePlanilhaGeral.R
\name{IndicePlanilhaGeral}
\alias{IndicePlanilhaGeral}
\title{IndicePlanilhaGeral}
\usage{
IndicePlanilhaGeral(
  wb,
  logo_innovare,
  nome_aba,
  pasta_nome,
  manter_ao_remover_dados = NULL,
  remover_dos_indices = NULL,
  indice_sigla_preco = "PR1",
  nota_roda_pe = NULL,
  inverter_condicional = c("IIQP", "IIC", "IICP", "IICP_BR"),
  espanhol = FALSE,
  anos,
  dados,
  Tab_Fundo_cinza,
  referencia_filtrado,
  print = FALSE
)
}
\arguments{
\item{wb}{Workbook onde a aba será criada.}

\item{logo_innovare}{Caminho do arquivo da imagem de logotipo a ser inserido na planilha.}

\item{nome_aba}{Nome da aba a ser adicionada ao workbook.}

\item{pasta_nome}{Nome da pasta exibida no cabeçalho.}

\item{manter_ao_remover_dados}{Colunas a serem preservadas ao remover dados de índices específicos.}

\item{remover_dos_indices}{Vetor de índices que devem ter seus valores removidos da planilha.}

\item{indice_sigla_preco}{Identifica o índice de preço e ajusta o valor "PR1" para "PR".}

\item{nota_roda_pe}{Vetor de notas de rodapé opcionais.}

\item{inverter_condicional}{Vetor de índices que invertem a condição de formatação condicional.}

\item{espanhol}{Se TRUE, exibe cabeçalho e títulos em espanhol.}

\item{anos}{Vetor de anos a ser exibido nos cabeçalhos.}

\item{dados}{Data frame com os dados para a planilha.}

\item{Tab_Fundo_cinza}{Data frame que define cores e estilos de fundo para linhas específicas.}

\item{referencia_filtrado}{Data frame com cores para formatação de fundo e texto para cada índice.}

\item{print}{Se TRUE, abre a planilha ao final.}
}
\value{
Workbook atualizado com a nova aba formatada.
}
\description{
Gera uma planilha detalhada e formatada em `wb` com uma estrutura de índice geral. A função cria uma aba com formatação específica e personalizada, incluindo estilos e mesclagens de células, baseados nos parâmetros fornecidos.
}
\details{
A função gera uma nova aba com layout e formatação específicos usando `openxlsx`, com inserção de dados de texto e numéricos e personalização avançada de células. O layout é otimizado para índices gerais, incluindo estilos com base nas cores dos indicadores e condições visuais para valores.

A função permite ajustar colunas específicas, configurar o cabeçalho em espanhol, inserir um logotipo e adicionar notas de rodapé. Suporta formatação condicional com cor de célula (vermelho/verde) com base em valores diferenciais e margem de erro para os índices especificados.
}
\examples{

base::print("Sem Exemplo")

}
