% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicePlanilhaImportancia.R
\name{IndicePlanilhaImportancia}
\alias{IndicePlanilhaImportancia}
\title{IndicePlanilhaImportancia}
\usage{
IndicePlanilhaImportancia(
  wb,
  logo_innovare,
  nome_aba,
  pasta_nome,
  dados,
  df_me,
  indice_sigla_preco = "PR1",
  nota_roda_pe = NULL,
  espanhol = FALSE,
  Tab_Fundo_cinza,
  referencia_filtrado,
  print = FALSE
)
}
\arguments{
\item{wb}{Workbook criado com a biblioteca `openxlsx`, onde a aba será adicionada.}

\item{logo_innovare}{Caminho para o logo da empresa, que será inserido na aba.}

\item{nome_aba}{Nome da aba dentro do workbook para a qual a planilha será gerada.}

\item{pasta_nome}{Nome da pasta de origem dos dados, exibido no cabeçalho da aba.}

\item{dados}{Data frame contendo os dados de índice para a planilha. Deve incluir colunas para índices e atributos.}

\item{df_me}{Data frame com dados de margem de erro associados aos índices.}

\item{indice_sigla_preco}{(Opcional) Sigla para identificar preços na coluna `IDAR`, padrão é `"PR1"`.}

\item{nota_roda_pe}{(Opcional) Vetor de caracteres para notas de rodapé a serem incluídas na aba.}

\item{espanhol}{(Opcional) Booleano para definir o idioma, `FALSE` exibe o cabeçalho em português, `TRUE` em espanhol.}

\item{Tab_Fundo_cinza}{Data frame que identifica linhas específicas para preenchimento de fundo cinza.}

\item{referencia_filtrado}{Data frame de referência para personalização de cores de fundo e texto na planilha.}

\item{print}{(Opcional) Booleano para abrir automaticamente o workbook após a execução da função; padrão é `FALSE`.}
}
\value{
Retorna o workbook com a aba recém-adicionada e formatada.
}
\description{
Gera e formata uma aba de planilha para exibir índices de importância associados a dados específicos.
Esta função configura uma aba dentro de um workbook, aplica formatações personalizadas para texto e dados numéricos,
e insere estilos e imagens conforme especificado. Opções de customização permitem incluir rodapés, títulos, e adaptar o idioma.
}
\details{
A função `IndicePlanilhaImportancia` organiza e formata os dados de importância dos índices dentro de uma nova aba de planilha.
A formatação inclui configurações específicas para cada célula e linha, permitindo um design claro e estruturado.
A função primeiro cria uma aba e escreve os dados de texto e numéricos, aplicando formatações de estilo com base nos
parâmetros de cor e formato fornecidos. Alturas e larguras de linhas e colunas são ajustadas automaticamente para acomodar
o conteúdo. Também são aplicados estilos diferenciados, como negrito e alinhamento, além de formatação de células baseada
na presença de indicadores específicos nos dados.
}
\section{Estrutura}{

- **Configuração de Alturas e Larguras**: Alturas de linha e larguras de coluna são definidas para otimizar a exibição dos dados.
- **Cabeçalhos e Títulos**: Insere um cabeçalho com o título "PLANILHA DE ÍNDICES" ou "PLANILLA DE ÍNDICES", baseado no idioma.
- **Formatos de Célula e Estilos**: Células recebem cores de fundo e estilo (negrito, alinhamento, etc.) com base nos valores de referência fornecidos.
- **Configuração de Margens de Erro**: Valores de margem de erro são ajustados conforme dados disponíveis em `df_me`.
- **Opções de Rodapé**: Notas de rodapé personalizadas podem ser adicionadas, exibindo informações adicionais sobre os dados da planilha.
}

\examples{

base::print("Sem Exemplo")

}
