% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicePlanilhaSatisfacao.R
\name{IndicePlanilhaSatisfacao}
\alias{IndicePlanilhaSatisfacao}
\title{IndicePlanilhaSatisfacao}
\usage{
IndicePlanilhaSatisfacao(
  wb,
  logo_innovare,
  nome_aba,
  pasta_nome,
  manter_ao_remover_dados = "Índice",
  remover_dos_indices = c("ISQP", "IEQP", "IIQP", "ISCP", "ISC", "IESC", "IIC", "ISCAL",
    "IECP", "IICP"),
  indice_sigla_preco = "PR1",
  nota_roda_pe = NULL,
  ano_titulo_planilha,
  espanhol = FALSE,
  dados,
  Tab_Fundo_cinza,
  referencia_filtrado,
  print = FALSE
)
}
\arguments{
\item{wb}{Workbook aberto onde a aba será criada.}

\item{logo_innovare}{Caminho para o arquivo de imagem do logotipo a ser inserido na aba.}

\item{nome_aba}{Nome da aba a ser criada no workbook.}

\item{pasta_nome}{Nome a ser exibido na planilha, indicando a pasta de dados usada.}

\item{manter_ao_remover_dados}{Colunas que permanecerão quando dados de determinados índices forem removidos.}

\item{remover_dos_indices}{Vetor com as siglas dos índices que devem ser excluídos dos dados.}

\item{indice_sigla_preco}{Sigla do índice de preço, usada para identificar linhas específicas.}

\item{nota_roda_pe}{Vetor com notas de rodapé opcionais a serem exibidas abaixo da tabela.}

\item{ano_titulo_planilha}{Ano ou título para a planilha, exibido no cabeçalho.}

\item{espanhol}{Lógico. Quando TRUE, exibe o cabeçalho em espanhol.}

\item{dados}{Data frame contendo os dados a serem exibidos na planilha.}

\item{Tab_Fundo_cinza}{Data frame para definir linhas que terão fundo cinza na tabela.}

\item{referencia_filtrado}{Data frame que contém as cores de fundo e texto para cada índice.}

\item{print}{Lógico. Quando TRUE, a planilha é aberta automaticamente após a criação.}
}
\value{
Retorna o workbook `wb` com as modificações aplicadas.
}
\description{
Gera uma planilha de satisfação formatada com cabeçalhos, estilos e dados, incluindo
logotipo, coloração, e agrupamento de células para indicadores específicos.
}
\details{
Esta função organiza e estiliza dados de satisfação em um workbook `wb` usando o pacote `openxlsx`.
Ela cria uma nova aba com cabeçalhos customizados e aplica formatação avançada de acordo com as especificações dos índices.
Estilos específicos são aplicados a células de acordo com as cores de cada índice,
as células são mescladas em determinadas posições, e uma imagem de logotipo é adicionada à planilha.

O código estrutura e formata a planilha de maneira a destacar informações importantes,
com ajuste automático das alturas das linhas, e inserção de notas de rodapé, se especificadas.
}
\examples{

base::print("Sem Exemplo")

}
