% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicePlanilhaSubs.R
\name{IndicePlanilhaSubs}
\alias{IndicePlanilhaSubs}
\title{IndicePlanilhaSubs}
\usage{
IndicePlanilhaSubs(
  wb,
  logo_innovare,
  nome_aba,
  pasta_nome,
  manter_ao_remover_dados = NULL,
  remover_dos_indices = NULL,
  indice_sigla_preco = "PR1",
  nota_roda_pe = NULL,
  inverter_condicional = c("IIQP", "IIC", "IICP", "IICP_BR"),
  espanhol = FALSE,
  anos,
  dados,
  Tab_Fundo_cinza,
  referencia_filtrado,
  print = FALSE
)
}
\arguments{
\item{wb}{Workbook no qual a aba será criada.}

\item{logo_innovare}{Caminho da imagem de logotipo a ser inserido na planilha.}

\item{nome_aba}{Nome da aba a ser criada.}

\item{pasta_nome}{Nome da pasta exibida no cabeçalho da aba.}

\item{manter_ao_remover_dados}{Colunas que devem ser preservadas ao remover dados.}

\item{remover_dos_indices}{Vetor de índices para remoção de dados específicos.}

\item{indice_sigla_preco}{Sigla do índice de preço que será ajustada para "PR".}

\item{nota_roda_pe}{Vetor de notas de rodapé a serem adicionadas (opcional).}

\item{inverter_condicional}{Índices que invertem a lógica de formatação condicional.}

\item{espanhol}{Lógico, define o idioma do cabeçalho (padrão: FALSE).}

\item{anos}{Vetor com anos a serem incluídos nos cabeçalhos da planilha.}

\item{dados}{Data frame com os dados a serem exibidos na planilha.}

\item{Tab_Fundo_cinza}{Data frame que define o estilo de fundo para linhas específicas.}

\item{referencia_filtrado}{Data frame com as cores dos indicadores para formatação.}

\item{print}{Lógico, se TRUE, a planilha é aberta automaticamente.}
}
\value{
Retorna o workbook `wb` com a nova aba formatada.
}
\description{
Gera uma aba em um `workbook` com formatação específica para índices de substituição, aplicando cores, mesclagem de células e formatação condicional de acordo com os parâmetros fornecidos.
}
\details{
A função cria uma nova aba com layout e formatação detalhada. Aplica-se um estilo específico para índices de substituição, com mesclagem de células e aplicação de cores específicas com base nos parâmetros de formatação e valores dos dados.
}
\examples{

base::print("Sem Exemplo")

}
