% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndiceVerificandoIndices.R
\name{IndiceVerificandoIndices}
\alias{IndiceVerificandoIndices}
\title{IndiceVerificandoIndices}
\usage{
IndiceVerificandoIndices(
  resultado_final,
  serie_historica_servidor,
  tabela_split,
  caminho_temporario = "99.temp cal erro",
  caminho_resultado = "Erros"
)
}
\arguments{
\item{resultado_final}{Data frame contendo os resultados de índices calculados localmente.}

\item{serie_historica_servidor}{Data frame com os valores de referência para a série histórica de índices, obtidos do servidor.}

\item{tabela_split}{Data frame com os dados 'dono_id', 'split_nome' e 'pasta_nome'.}

\item{caminho_temporario}{(Opcional) String com o caminho onde os arquivos temporários serão armazenados durante a execução. Padrão: `"99.temp cal erro"`.}

\item{caminho_resultado}{String com o caminho onde os arquivos de erro finais serão armazenados. Padrão: `"Erros"`.}
}
\value{
A função não retorna um objeto em R. Em vez disso, salva arquivos RDS com logs de erros em `caminho_resultado` e apaga
o conteúdo do diretório temporário.
}
\description{
Esta função compara resultados de índices calculados localmente com uma série histórica do servidor, verificando discrepâncias
e salvando logs de erro de forma organizada em arquivos temporários e permanentes.
}
\details{
A função `verificando_indices` organiza e verifica possíveis discrepâncias entre os dados calculados localmente e a série histórica
do servidor em várias dimensões, incluindo número de linhas, amostras pretendidas e comparações de variáveis de importância.
Discrepâncias são identificadas e armazenadas como objetos RDS em subdiretórios temporários específicos.
Cada erro detectado (como número de linhas incorreto ou discrepâncias em importância) é salvo em um arquivo RDS temporário para posterior
agregação em um local de armazenamento permanente.
}
\section{Etapas de Verificação}{

- **Configuração de Diretórios Temporários**: Cria e limpa subdiretórios para armazenamento temporário de logs de erro.
- **Verificação por Grupo**: Processa cada grupo (`dono_id`) em `resultado_final` e verifica o número de linhas e dados de amostra,
  comparando-os com a série histórica do servidor.
- **Comparação de Índices**: Compara índices calculados localmente e armazenados no servidor e identifica discrepâncias em
  variáveis de `media`, `indice`, e `importancia`.
- **Salvamento de Resultados**: Logs de erro consolidados são salvos em arquivos RDS finais e organizados por tipo no
  diretório especificado.
}

\examples{

base::print("Sem Exemplo")

}
